create scripts to generate icons for vite and for webpack
